# chore: micro optimization