@Cadienvan 2025-03-24
Thanks @Eomm for the awesome work! I fixed the types and added excludedRoutes checks on the onRoute, nice tip!!